changed neighbor index types to unsigned integers

+++ sph/include/sph/update_h.hpp
@@ -6,20 +6,20 @@ namespace sph
 {
 
 template<class Dataset>
-void updateSmoothingLength(size_t startIndex, size_t endIndex, Dataset& d, size_t ng0)
+void updateSmoothingLength(size_t startIndex, size_t endIndex, Dataset& d, unsigned ng0)
 {
     using T     = typename Dataset::RealType;
     const T c0  = 7.0;
     const T exp = 1.0 / 3.0;
 
-    const int* neighborsCount = d.nc.data();
-    T*         h              = d.h.data();
+    const unsigned* neighborsCount = d.nc.data();
+    T*              h              = d.h.data();
 
 #pragma omp parallel for schedule(static)
     for (size_t i = startIndex; i < endIndex; i++)
     {
-        int nn = neighborsCount[i];
-        h[i]   = h[i] * 0.5 * pow((1.0 + c0 * ng0 / nn), exp);
+        unsigned nn = neighborsCount[i];
+        h[i]        = h[i] * 0.5 * pow((1.0 + c0 * ng0 / nn), exp);
 
 #ifndef NDEBUG
         if (std::isinf(h[i]) || std::isnan(h[i])) printf("ERROR::h(%lu) ngi %d h %f\n", i, nn, h[i]);
